[FEAT] Adds fields of views in the state machine
It is a first way to describe which objects are visible to objects declared as
"viewers". These objects are then stored in fov, then in a text file, that is 
given to the client.

TODO : instead of passing by a text file, the changes in fov will soon have to 
be transmitted by callbacks (once the fusion between server and  state 
machine is done).