[FIX] Timers are now working with signals like other actions
Timers were stored in a global variable, they are now attributes of the model
which let them being shared across threads and signals firings.
The system relied on is_alive() function to recognize new threads/timers
from old ones. It now relies on a boolean associated to each timer, which is
set to False when the run() method terminates.

Fixes #5, fixes #1