add convert_csvfile
Measurements should refer to an image id in order to be inserted in the
database (by insert_csvfile). CSV generally refer to subjects (in their
first column). Hence this function is able to fetch the image_id for
every subject (provided each subject has only one image available) and
convert a CSV to this database-compliant format.

Still the header of the CSV should be :
header = ['subject', 'structure', 'side', 'measurement', 'unit',
'value', 'software', 'comments']